Split NuGet source policy by owner
Move source options, source resolution, and package search into files named for their current responsibilities. Remove the unused raw search-response compatibility entry point.
